#427 Right-Anpassung: Nur eigene statische Seiten bearbeitbar machen [erledigt] Außerdem weitere kleinere Änderungen, Bugfixes und Anpassungen

+++ _ajax.php
@@ -1,31 +1,36 @@
 <?PHP
 /* 
-	01-Artikelsystem Copyright 2006-2008 by Michael Lorer - 01-Scripts.de
+	01-Artikelsystem Copyright 2006-2011 by Michael Lorer - 01-Scripts.de
 	Lizenz: Creative-Commons: Namensnennung-Keine kommerzielle Nutzung-Weitergabe unter gleichen Bedingungen 3.0 Deutschland
 	Weitere Lizenzinformationen unter: http://www.01-scripts.de/lizenz.php
 	
 	Modul:		01article
 	Dateiinfo: 	Bearbeitung von eingehenden Ajax-Requests
-	#fv.3010#
+	#fv.310#
 */
 
+// Artikel / Seiten löschen
 if(isset($_REQUEST['ajaxaction']) && $_REQUEST['ajaxaction'] == "delarticle" &&
    isset($_REQUEST['id']) && !empty($_REQUEST['id']) && is_numeric($_REQUEST['id'])){
-    if($_REQUEST['static'] == 0 && $userdata['editarticle'] == 2){
-			mysql_query("DELETE FROM ".$mysql_tables['artikel']." WHERE id='".mysql_real_escape_string($_REQUEST['id'])."' AND static = '0' LIMIT 1");
-			delComments($_REQUEST['id']);
-			}
-		elseif($_REQUEST['static'] == 0 && $userdata['editarticle'] == 1){
-			mysql_query("DELETE FROM ".$mysql_tables['artikel']." WHERE id='".mysql_real_escape_string($_REQUEST['id'])."' AND uid = '".$userdata['id']."' AND static = '0' LIMIT 1");
-			delComments($_REQUEST['id']);
-			}
-		elseif($_REQUEST['static'] == 1 && $userdata['staticarticle'] == 1){
-			mysql_query("DELETE FROM ".$mysql_tables['artikel']." WHERE id='".mysql_real_escape_string($_REQUEST['id'])."' AND static = '1' LIMIT 1");
-			delComments($_REQUEST['id']);
-			}
 	
-	echo "<script type=\"text/javascript\"> Success_delfade('id".$_REQUEST['id']."'); </script>";
+	    if($_REQUEST['static'] == 0 && $userdata['editarticle'] == 2)
+		mysql_query("DELETE FROM ".$mysql_tables['artikel']." WHERE id='".mysql_real_escape_string($_REQUEST['id'])."' AND static = '0' LIMIT 1");
+	elseif($_REQUEST['static'] == 0 && $userdata['editarticle'] == 1)
+		mysql_query("DELETE FROM ".$mysql_tables['artikel']." WHERE id='".mysql_real_escape_string($_REQUEST['id'])."' AND uid = '".$userdata['id']."' AND static = '0' LIMIT 1");
+	elseif($_REQUEST['static'] == 1 && $userdata['staticarticle'] == 2)
+		mysql_query("DELETE FROM ".$mysql_tables['artikel']." WHERE id='".mysql_real_escape_string($_REQUEST['id'])."' AND static = '1' LIMIT 1");
+	elseif($flag_static == 1 && $userdata['staticarticle'] == 1)
+		mysql_query("DELETE FROM ".$mysql_tables['artikel']." WHERE id='".mysql_real_escape_string($_REQUEST['id'])."' AND uid = '".$userdata['id']."' AND static = '1' LIMIT 1");
+			
+	if(mysql_affected_rows() == 1){
+		delComments($_REQUEST['id']);
+		echo "<script type=\"text/javascript\"> Success_delfade('id".$_REQUEST['id']."'); </script>";
+		}
+	else
+		echo "<script type=\"text/javascript\"> Failed_delfade(); </script>";
 	}
+
+// Kategorien löschen
 elseif(isset($_REQUEST['ajaxaction']) && $_REQUEST['ajaxaction'] == "delcat" &&
    isset($_REQUEST['id']) && !empty($_REQUEST['id']) && is_numeric($_REQUEST['id'])){
 	$sqlcat = "SELECT id,catpic FROM ".$mysql_tables['cats']." WHERE id = '".mysql_real_escape_string($_REQUEST['id'])."'";
@@ -52,6 +57,4 @@ elseif(isset($_REQUEST['ajaxaction']) && $_REQUEST['ajaxaction'] == "delcat" &&
 else
 	echo "<script type=\"text/javascript\"> Failed_delfade(); </script>";
 	
-	
-// 01-Artikelsystem Copyright 2006-2008 by Michael Lorer - 01-Scripts.de
 ?>
